requirements.txt
Django==3.1.7
djangorestframework==3.12.2
django-filter==2.4.0
drf-yasg==1.20.0
psycopg2==2.8.4
PyJWT==1.6.4
django-safedelete==0.5.5
django-storages==1.9.1
boto3==1.12.6
python-dateutil==2.8.1
celery==4.4.2
sentry-sdk==0.17.0
babel==2.8.0
requests==2.23.0
urllib3==1.25.11
chardet==3.0.4
apns2==0.7.2
ipython==7.16.1
redis==3.2.0
django-celery-results==2.0.1
drf-extensions==0.2
django_redis
mysqlclient==2.1.0